Qeueing Aris Status Request and Others Job

+++ database/migrations/2024_12_07_055108_add_image_column_to_how_to_threads_table.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('how_to_threads', function (Blueprint $table) {
-            $table->dropColumn('image');
+            if (Schema::hasColumn('how_to_threads', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
 };
